requirements.txt
torch
gradio
torchvision
numpy
Pillow
matplotlib
